backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
python-dotenv
requests
passlib
